kill -9 the ruby depselector on timeout
Timeouts on the ruby depselector have recently caused
zombie ruby processes to stick around that are not
able to tell that stdout has been closed. This patch
captures the pid of the ruby process at startup
and then sends a kill -9 to the pid when the depselector
process returns a timeout response.