Restrict darklaunch data insertion to only run on the live redis host
We initialize darklaunch data for redis in the redis lb cookbook.

Redis is only available on a single host, and while we could remotely
access it, it doesn't make sense to update it more than once anyways.

Eventually we should consider making this migration instead, since it
probably shouldn't be done on every reconfigure.