Revert "Revert "Convert file IO to raw mode""
This reverts commit d6b935d34c4210d9e9450cb0749716e4e4ff44ae.

This restores the usage of raw IO mode. We had seen earlier failures
of bookshelf under load when running in raw mode, and had
(incorrectly) blamed this patch for the problem.

Further research shows that the load failures were due to logging
backing up. Enabling raw mode appears to have sped up the rest of the
system, causing the logging problem to present sooner.