Increase async worker thread pool size to 10
Increasing the number of async worker threads from the defaut of 0
increases the performanc of IO-heavy apps like bookshelf.

Load testing different configurations of +A resulted in a large
performance increase from 0 to 8 workers, and no appreciable
preformance increase form 8 to 16. The value 10 was chosen as
it is the default in Erlang R16B.