do not check for banned/whitelisted IPs in EC
EC does not expose controls for the configuratoin of banned and
whitelisted IPs. In addition, the checks in place are not ipv6
compatible.

If this functionality is determined necessary in future release, it will
be done in a way that integrates the native nginx address range
checking instead of custom matching.