Separate redis connect and keeaplive timeouts and change defaults
Previously redis_connection_timeout was being used for both the
timeout for connect() operations and as the keepalive_timeout which
determines how long connections stay in the pool.

This commit separates these two timeouts allowing them to be
configured independently.

Further, since both timeouts are in milliseconds, I have increased
their default values to 1 second for connect() and 2 seconds for
keepalive.