[OC-3107] Do not run compaction on backup backend
Previously we checked for the existence of the _users.couch database
to avoid running compaction on the backup backend.  Unfortunately, if
couchdb ever starts up on the backend when it is still the passive
(say during a reboot), this database may get created.

This patch modifies the compaction script to check if it is running on
the backend and exit if that is the case.  It also removes the
previous check from the cron entry.