Add users container type, and use opscode_account DB when no org present.
Add support for fetching global users container from couch. This is
triggered based on no org being present - if no org is present, then the
referenced object type must be not belong to an org.  (Or if this is not
true, it should fail based on its absence in global containers.)

This allows oc_chef_authz to be aware of global group containers, such
as for users.