Allow a custom script to be used for IP/block device failover
This is straight from Irving's ec-metal repo and is a relatively small change
to cluster.sh that checks to see if a given script exists (or rather two
scripts, one for block device failover and another for IP address failover).
If neither of these scripts are present, then the behavior remains unchanged
and DRBD failover continues to be used.
If the scripts are  present, they are run with either 'attach' or 'detach' as a
parameter, and should perform the relevant failover commands as needed. An
example is in ec-metal for EC2 with EBS volumes and VPC Private IPs.

This will allow us to develop different HA failover methods independently,
just dropping the relevant script in place.