Update to TK and Omnibus 3.1
In order to get around some issues with the git-caching
version of omnibus-ruby (3.0+), we must pull in the
latest omnibus-software which contains fixes to the logrotate
bulid definition to ensure that logrotate is actually
installed.

TK and Berkshelf should be installed via ChefDK. These
tools don't actually work with vagrant from within the
bundled install, and they slow down the build when
gecode and ridley and Berkshelf need to get installed
before executing a build that doesn't even leverage those
tools.