use berkshelf2 software config
The berkshelf software config drifted and is catered to
installing berkshelf from master and at versions > 3. We're
still using berkshelf 2 here, and as such we don't need it
to depend on gecode. I've ported the pre-berks-3 version
of the berkshelf software config and named it berkshelf2.

This should fix the failing builds on centos5.