Remove authz_id from insert_cookbook_versions WHERE clause
The authz_id is not necessary to guarantee a unique result as the name
and org_id combination is also constrained to be uniq.  Further, when
new cookbooks concurrently, it is possible that the authz_id we have
is incorrect, leading to a failure.

Since we depend on the record order, we have to do some ugliness to
remove authz_id from the arguments used.