Change how custom ha-status is done
Previously the custom status command would output 'master' or 'backup' and the
ha-status command would compare it to the current HA status, reporting an
error if there was an inconsistency. This changes it to simply check to see if
the script returns a non-zero status or not, and passes in the current status
for the script to check itself.

The reason for this is that the DRBD status script needs to do more than just
output if we are master or backup, and does some other sanity checks.