Fix for upgrades the bad rabbit wait script.
The script was checked in with a copy-paste error into OSC
where it checks that EC rabbit is running. This wasn't noticed
before b/c it has a bail option where it does nothing if EC isn't on the
box. EC was of course never on an OSC box, until now.

This fix just makes the script a blank file, so upgrades can progress
and be rerun as needed. Further details are in a comment in the code.