fix setup and cleanup for association tests
When adding users as actors directly to the admins and users group,
they are not cleaned up when dissociating user from said org, which
only deletes the USAG for the user (causing cascading authz cleanup).

This commit will fix the issue because we are actually deleting the
user after the test, which will also cause cascading authz cleanup.

There should be further work to actually remove the actor from the
users group before the dissociation.