Make upgrade code more module
This involves turning the osc upgrade portion into a class
and breaking the upgrade into three main components:

download
transform
upload

This then enables these parts to be reused in other commands
by instantiating the class and calling those parts that are needed.

There is a bit of method missing magic to catch calls to methods
that exist in the omnibus-ctl context outside of the osc upgrade class.
This works by passing in teh omnibus-ctl context and calling it as
needed.

The references to OSC and EC in the log messages given to the user
have also been scrubbed and replaced with Chef 11 and Chef 12
respectively, as this will make more sense to the user.