Fixed a race condition that prevents the gen_fsm from transitioning from the halting state to the ready state.
The conditions for this bug to be triggered are an initial size of zero things to migrate and mover_manager:status() being called in a loop.

Transition from halting state to ready state is dependant on a timeout occuring. Any state transition (which includes transitioning from a state to itself), resets the gen_fsm timeout. The status call was clearing the gen_fsm timeout, since it had a non-zero timeout arg. This would prevent a proper transition from the halting state to the ready state if status was called in a loop.

The solution is to give the status operation a timeout arg of zero when state == halting so that it could not reset the timeout, allowing the halting -> ready timeout to occur.