Get rid of pre-created org deletion and log cleaning post 11.1-stable.
We don't need to delete pre-created orgs in upgrades past where 11.1-stable is since couchdb will not exist there. Also, let's not clean up the mover logs in the same upgrades since we will want to be able to see any failures in this set of migrations that will occur at once.