Upgrade: Make chef server url and Chef 11 config options fully configurable
This commit splits the chef server url into two options, one for
Chef 11 and one for Chef 12, since the may not be the same if the
upgrade is not being done on the same box.

It also adds options to specify the admin user and user key to use
during the Chef 11 upgrade, in case the defaults have been modified.