Make the data storage directories user configurable during upgrade
Previous in the main upgrade command these directories were randomly
generated, but now they can be specified by the user if desired.

This enables a user to script their creation and deletion after an
upgrade process if so desired, so they will also be in a predictable
place of the user's choosing.

This also updates the individual commands to reflect the change in
upgrade option naming that this additions to the main upgrade commands
caused, as there were some option naming collisions for the short names.