Add msg about upgrade data and packages still on disk
After an open source Chef 11 to Chef 12 upgrade, the data
download and transformed from the upgrade process will still be on
disk. Give the user a msg this is the case and inform them it can be
backed up or remove.

Also note that the Chef 11 package is still on the system and can
now be safely uninstalled.