opscode-reporting chef-server-running back-compat
Because we've moved the opscode-solr resource to opscode-solr4,
opscode-reporting can no longer find the correct configuration
values in the chef-server-running json file when tyring to configure
the search attributes.

This fixes by re-adding the ospcode-solr key that is a copy of the
newly-named opscode-solr4. This can be removed when reporting is
updated to use the new fields.