fix post-12 cleanup for HA installs
The previous version of the post-12 cleanup did not take into account
that data directories for databases are in a different location for
HA installs. This change re-populates that data in the
chef-server-running config, and fixes an issue where said config
was storing all of the data for opscode-solr as opscode-solr4, which
would have caused a cleanup to delete all of the opscode-solr4 data.