move the addon install before bootstrap
If the addon install fails for some reason, we'd like the
system to be in a state where we can re-run the reconfigure
and it will pick up where it left off. By installing the
addons after the bootstrap data has been provisioned but before
we'd written the bootstrap file, we were in a situation
that subsequent reconfigures would fail if the package
install failed for any reason. This fixes that.