turn off addon installs by default
Due to the variance in speed for addon installs, we've elected to
turn off the feature that installs them by defaut for new installs.
The addons were taking between 5 and 50 minutes to install, and
this experience is less than delightful if all you'd like to do
is quickly get a Chef Server up and running.