Use the X-Forwarded-For header in nginx logs instead of the remote address, based on the value of node['private_chef']['nginx']['log_x_forwarded_for'].  This attribute defaults to true if the topology is set to 'ha' or 'tier' and false otherwise.
This change makes the nginx logs contain the address of the requesting client rather than
the socket address of the remote node.  When the frontends are behind a load balancer, the
remote node is always the load balancer, but most load balancers populate the X-Forwarded-For
header with the IP address of the remote node that made the original request.