lock_deps: don't ignore meck
Ignoring meck when locking deps is causing us to lock to the esl/meck
repo instead of eproxus/meck. When we build, we're actually checking
out eproxus/meck and compiling againts it, so we end up telling the
`make rel` command to pull in and compile something that we haven't
tested against. Since esl/meck is now deleted, `make rel` fails.