pull in meck from eproxus
The esl/meck project has gone away, so we need to be using the meck
project from the eproxus repo. The previous version that was pinned
in the esl project also no longer exists, so I also bumped this to
the SHA of master on the eproxus repo. Since meck is used only in
testing, this will not effect the running application, and tests
still pass.