Revert "Merge pull request #559 from opscode/of/oc_erchef_in_11_stable"
This reverts commit 806e6d13f00c15a390af63e98e4d09f13e826cb5, reversing
changes made to 08098a59805b6f87b9e58d1f4eca58bbe1a3c317.

This commit rolls back the Erchef API to a version that defaults
darklaunch features to CouchDB mode. This behavior is needed for
existing enterprise addons that expect to be able to make API requests
directly to the Erchef service, and not through the Nginx reverse
proxy, which sets correct darklaunch headers.

Further work will need to be done to back-port the latest Erchef
service and retain this default behavior.