Refactored superuser bootstrap process.
Updated bootstrap in a variety of ways. Most importantly, pulled in chef-server-bootstrap to replace opscode-test. This allows us to get rid of a variety of now deprecated ruby code, including:
+ opscode-test
+ opscode-billing
+ couchrest
+ opscode-shared
+ mixlib-* (still used by client)

It also removed inserting the pivotal password on bootstrap, as this was pointless, since it isn't used internally, and the unhashed password is saved nowhere, so a user would have to regenerate the password anyway to use it.