Pull down gpg key used to verify add ons as part of the build process.
Some versions of RHEL do not support the proper ssl protocol in yum (SNI) to
grab the gpg key from the remote that it is currently stored in. Pull it
down as part of the build process and place in on disk in the cache path
so that RHEL systems can have it locally to verify when downloading add ons.