fail on org creation conflict during setup
If we configure pedant to create an organization, then we
expect the organization to not exist when starting the tests.
This commit removes the logic to reassign the validator key
(a destructive operation) if the org already exists. To use
an existing org, simply do not set the :create_me flag in the
pedant config.