mover_manager: always reset 0 timeout when calling status
Because of the way we're triggering state transitions within the
mover_manager gen_fsm (via 0ms timeouts), calling mover_manager:status
has shown the ability to prevent a state transiation and cause
migrations to hang. This commit ALWAYS resets the gen_fsm timeout
value to 0 when calling mover_manager:status, so that we no longer
run into this situation.