Do not hardcode superuser name
Previously, we had hard-coded a check for the superuser default name
<<"pivotal">> in chef_sql:is_user_in_org, because superusers are by
default in all organizations for purposes of permissions.

However, we already had the logic to do this in a configurable way
in oc_chef_wm_base - we just had it disabled by default, and were
instead relying on the hard-coded check in chef_sql.  This change
uses that existing functionality, and removes the hard-coded check in
chef_sql:is_user_in_org.

Note that this only bypasses org membership check, and not verification
of the requestor. (So we still verify that superuser is who it says it
is.)