mover_manager: correctly handle start messages in non-ready state
Some time ago there was a change in the message format for start
messages in the mover_manager gen_fsm state machine. That change
didn't get propagated to all of the fun heads, and can cause
crashes when we simply would prefer to ignore the messages. This
commit fixes that error.