For client/user GET se public_key & version from keys table
WHen GETing a client or user, retrieve the public_key and
public_key_version comes from the keys table. This ensures
that the correct value ('default' key for the object) is returned,
instead fo the potentiall outdated value in the/user table

If no 'default' key is present for the client or user, the returned JSON
will be null.