Allow multiple LoginAttr fields in LDAP user record.
It is not uncommone for an LDAP record to have multiple copies of a
field, allowing users to be found more easily during search and to
account for name changes through time.  This change accounts for
multiple copies of the LoginAttr field by simply taking the first one
that matches the username we were given.

Closes chef/chef-server#71