First pass at adding a new cookbook_artifacts endpoint
The intent is to use this endpoint for policy files.
The new endpoint as of now consists of:
```
GET/POST BASE_URL/cookbook_artifacts/NAME/IDENTIFIER
```

It is based off this RFC:
https://github.com/chef/chef-rfc/blob/fef5ee5e7c8c051948cde51f49e96cb02a20d5b4/rfc022-arbitrary-cookbook-identifiers.md

Basic idea is to have a new cookbook endpoint that behaves
largely in the same way as the old cookbooks endpoint, except
it distingues versions based on arbitrary identifiers,
and no longer on version numbers.

This new endpoint's data is stored entierly separately from
the old cookbooks endpoint's one.

The implementation is somewhat different from the one of the
old endpoint though, in that that is hides (most of) the nitty
gritty details away in the DB, which allows for a cleaner code
on the erlang side.

Schema modifications:
https://github.com/chef/enterprise-chef-server-schema/pull/25

Added CT tests on this new endpoint.

Also some minor cleanups on the way.

Future work:
 * add more validation
 * add a DELETE method on this endpoint
 * add `GET BASE_URL/cookbook_artifacts/` and
   `GET BASE_URL/cookbook_artifacts/NAME` endpoints