Do not put the parsed date tuple into ejson
parse_binary_ejson is expected to return a valid ejson object,
but the value we get from converting the string time representation
to an erlang term is not valid ejson. This breaks analytics tracking of
field values, and is bad form anyway.

Updated so that we validate the field but do not capture the parsed value until
we're actually creating a record in new_record.