Import schema from enterprise-chef-server-schema
Enterprise-chef-server-schema often needs to change alongside the code
that makes use of the new schema.  Having the schema in a separate
repository introduced unncessary friction into the development
process.

deps/chef-server-schema is an import of chef-server-schema at 1.0.4

If you have stumbled upon this repository via git-log or git-blame and
need more history, see the following upstream repositories:

https://github.com/chef/enterprise-chef-server-schema

and

https://github.com/chef/chef-server-schema