Add supoort server-side key generation to keys REST API
Added support to the keys endpoints for server-side generation of a new
client key  Validates to ensure that either public_key or
create_key is provided, and will fail with a 400 if both are provided.

If a key is generated (PUT or POST body contains "create_key":true)
it will be returned as "private_key": "$KEY" in the response JSON. If it
is not requested, a valid public key must be sent and the private_key
field will not be sent in the response body.

This is implemented under API v0, since it was an additive change.

ChangeLog-Entry: add support for server-side generation of keys via the
keys API by accepting 'create_key': true in the request body.