Api v1 - clients, users complete implementation
ChangeLog-Entry: This introduces new, consistent handling of keys across
the clients, users, and keys endpoints.  It disables the updating of keys
via the clients and users endpoints - all keys are to be managed via the
keys endpoints after initial creation. This deprecates the previous key
creation and update, though it is still supported for backwards
compatibility.