dialyzer fixes for oc_chef_wm_org_invites
This commit reduces the number of dialyzer output lines to from 84 to
74:

```
> SKIP_DIALYZER=false make dialyze | wc -l
make: *** [dialyzer] Error 2
      74
```

This commit changes the input to chef_object:new_record from
oc_chef_wm_org_invites:create_association. We were previously calling
the function with AuthzId set to `{authz_id, UserId}`, but the
new_record function takes either an `object_id()` or the atom
`unset`.

This will result in a slight behavior change that needs testing.
The moser library calls chef_object:new_record for
oc_chef_org_user_invite with the `{authz_id, AuthzId}` pattern, so
we'll need to update that code as well to conform to the correct spec.
Fortunately, we should see dialyzer break when we pull in the new
dependencies there.