Implemented X-Ops-Server-API-Version HTTP header in server response.
ChangeLog-Entry:

In the response header, `X-Ops-Server-API-Version` will return json of the format:

```
{
  "min_version":"<non-negative integer>",
  "max_version":"<non-negative integer>",
  "request_version":"<integer>",
  "response_version":"<integer>"
}
```

Where:
+ `min_version` is the minimum API version the server supports.
+ `max_version` is the maximum API version the server supports.
+ `request_version` is an integer representing the desired API version from the client passed via X-Ops-Server-API-Version
defaulting to 0 if X-Ops-Server-API-Version was not sent by the client. If X-Ops-Server-API-Version sent by the client contained
an invalid value (not an integer), then -1 is returned in `request_version`.
+ `response_version` is an integer representing the API version used by the server to process the request.
It either matches what the client requested in X-Ops-Server-API-Version or is -1 if a 406 occurred (which happens
when X-Ops-Server-API-Version sent by the client was invalid or not in the supported range of the server).