Provide pooler bounded-queuing configuration for all pooler controlleed queues
Previously, the only connection pooler with user-controllable queuing settings was
the sqerl pool in oc_erchef.  Now the queueing can be added to the following pools:

- sqerl in oc_erchef,
- depsolvers in oc_erchef,
- oc_chef_authz in oc_erchef,
- sqerl in oc_bifrost

By default, queueing is diabled because the relevant pooler_timeouts
are set to 0.  The queue max for sqerl is left at 20, while all others
are set to 50 which is the existing default in pooler.

As mentioned queueing is only used when a timeout > 0 is given.  This
also requires that pooler:take_member/2 is used rather than
pooler:take_member/1.  Thus, one callsite of pooler:take_member/1 has
been converted.