Use custom context configuration for solr jetty
Previously, the default context configuration was setting
tempDirectory in the context configuration for solr. In Jetty 8, this
means that jetty will initially extract the WAR to the specified
directory and only re-extract it when the mtime on the WAR is newer
than the mtime on the directory.  Unfortunately, this leads to cases
where solr will not get updated on chef-server upgrades if the current
version was installed after the new version was built.

To avoid custom logic to update the WARs mtime, we configure jetty
without tempDirectory, leading it to extract the WAR on every restart.
To avoid depending on /tmp, we also set java.io.tmpdir.

Fixes #323