Update GET /policies response with new format+info
The old response format was a JSON Array of policy names, e.g.

```json
[ "foo", "bar" ]
```

In the ChefDK UI, we collate All policyfile revisions by policy_group
and optionally display "orphaned" policies, therefore the API must show
all of the revisions of a policy, not just the names. The new format is:

```json
{
  "policy_name": {
    "uri": "URI",
    "revisions": {
      "abc123": {},
      // ...
    }
  }
}
```