make postgres connections consistently use tcp everywhere possible
The goal here is to ensure common connection behavior for all actions that apply to
both external and managed postgres instances.  In the single case where we
need to connect via pipe instead of tcp (assigning password to the
opscode-postgres user in a managed installation), it gets made explicit
that we're using a local connection.