[oc_bifrost] Concurrency fix for update_acl
Added pg_advisory_xact_lock to lock the target id in update_acl

This will block concurrent updates to the target id using
update_acl, which avoids deadlocks seen in cases of multiple
updates to an actor when adding them to multiple groups with concurrent
calls. These updates will still complete, with the final update
being the one that is .retained

There are some things about this that make it less than ideal:

1. It doesn't address the root cause of why these requests were
   coming in milliseconds apart.

2. This is a blocking lock request. Rather than fail subsequent requests,
   this will allow them through, but it will block them so that they're
   processed first come first served.  In the event that we see a large
   number of concurrent requests this could cause a backlog of held
   connections in the bifrost pool.

   A better long-term solution here might be to make it non-blocking
   (pg_try_advisory_xact_lock) and handle the failure in bifrost, as
   well as handling the specific message in erchef so that we can
   provide meaningful feedback via the API: a second requests to update the
   same ACL is coming in at the same time.

   However, this occurs infrequently  enough that it may not be a
   concern.

3. Because update_acl actual replaces all acls, this means that When multiple
   requests come in to update the ACLs against the same target, the last
   processed will the the final ACL.  There are no guarantees as to
ordering, so it the last one processed may not be the last request made.

ChangeLog-Entry: This corrects a concurrency issue in bifrost where
multiple ACL updates to the same entity could cause a deadlock.
ChangeLog-Entry: This corrects a concurrency issue in bifrost where
multiple ACL updates to the same entity could cause a deadlock.