By default, bind oc_id to IPV4 address
When using localhost as the vip instead of 127.0.0.1 (as we do
elsewhere), we may end up binding oc_id on port 9090 of ::1 (IPV6).
This can cause nginx to get connection refused, because it will
resolve to 127.0.0.1, but that is not the same.

Setting the vip attribute to 127.0.0.1, AND using the -b option for
rails will ensure that this is set correctly to listen on the IPV4
address.

ChangeLog-Entry: Change oc-id vip back to 127.0.0.1 to avoid possible
error with nginx; add -b option for Rails and make vip fully
configurable so it can work properly in IPv4 and IPv6 environments