Fix invalid envy validations in chef_index
envy:get's validation logic accepts arrays, but expects the array to be
an array of other type specification. The previous code assumed that
envy would validate against the values in the array. To check if the
passed option is a member of an array, we use a new helper function
one_of/1 which returns a fun() which will do the validation.