Use spawn/3 to facilitate async reindexing in all index modes
In the batch and inline index modes chef_index:add does not return until
solr has returned a response. This is great for reindex operations as we
will end up reindexing one object at a time. This is a quick attempt to
provide a chef_index:add_async by wrapping chef_index:add/4 in spawn/3.