Use case-insensitive search for user lookup by external_auth_id
Almost all of the typical values of ldap['login_attribute'] are case
insensitive according to the LDAP schema. However, we previously used a
case-sensitive lookup when trying to find an existing user record.

This commit changes the default lookup to be case insensitive but
provides a user-controllable override for the rare case where users are
setting a case-sensitive login_attribute.

ChangeLog-Entry: Fix bug where logins via LDAP failed because of case
sensitivity.