Cloudsearch Support
This patchset adds support for using AWS CloudSearch as the search index
used by the Chef Server. CloudSearch provides a Solr-like API that can
provide a search service similar to that provided by the Chef Server's
managed version of Solr 4.

When using CloudSearch as a backend, the following limitations apply:

All searches are case-insensitive Search results are limited to the
first 10000 results. The rabbitmq search indexing queue cannot be
used. Only direct-from-erchef (inline or batch) modes are supported.
CloudSearch must be correctly configured by the user with the correct
index fields and access policies.