[oc_chef_wm] Return the correct "total" in search responses
Solr may return IDs that are no longer in our database. We filter such
IDs from the response to account for this. Unfortunately, the previous
code was returning the count of responses from Solr, which will be
incorrect if filtering has occurred. This patch updates the code to use
the count of responses from Postgresql (i.e the filtered count).