[oc-chef-pedant] Remove assertions on "total" field in search responses
The "total" field is not guaranteed to match the actual count of
returned objects.  In practice, this has been working for solr since
we force an explicit commit between deletes and creates. However,
these tests fail when run against CloudSearch which does not support
an existing commit.

NB: The total is reported directly from Solr/Cloudsearch and may
contain delete nodes that are filtered from the result. However, we
cannot adjust total or we risk clients hitting errors related to
pagination since Solr/Cloudsearch controls pagination.