Don't run eunit on mover's vendored code.
We don't modify this, and it's currently preserved from a fixed
version to ensure mover's back-compatibility.  Because it's a regular
source of intermittent CI failures that we can't address(because we
can't change the vendored code), this change explicitly tests
only the parts of mover that can change.

ChangeLog-Entry: [chef-mover] do not run intermittently failing tests on dependencies
that we can't change for backward-compatibility reasons.