update minimal cookbook versions query to force query plan
This query uses two CTEs (inputs and min_cve) to force the query planner
to first join cookbook_versions (min, maj, patch) with cookbooks (id, name)
before filtering with inputs.

Without the min_cv CTE, the query planner would choose to join inputs with
cookbook_versions before filtering by name. In the worst case scenario (all
cookbooks have the same version number), the inputs and cookbook_versions
join behaved like a cartesian product of all the inputs and the cookbook
versions.

Here are some details on the worst case scenario mentioned in the PR:

* 201 cookbooks, `cookbook_(1..200)` and `master`
* `master` depends on all other cookbooks
* no other dependencies
* all cookbooks are at version `1.1.1`

Before this PR, this is what `explain analyze` on the minimal cookbook
versions query returns for this data set:

```
                                                              QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------
 Nested Loop  (cost=4.75..28.42 rows=1 width=578) (actual time=3.825..418.399 rows=201 loops=1)
   Join Filter: (inputs.name = c.name)
   Rows Removed by Join Filter: 40200
   CTE inputs
     ->  Function Scan on unnest  (cost=0.00..1.00 rows=100 width=56) (actual time=0.062..0.350 rows=201 loops=1)
   ->  Hash Join  (cost=3.75..27.03 rows=1 width=536) (actual time=1.541..68.215 rows=40401 loops=1)
         Hash Cond: ((v.major = inputs.major) AND (v.minor = inputs.minor) AND (v.patch = inputs.patch))
         ->  Seq Scan on cookbook_versions v  (cost=0.00..21.01 rows=201 width=528) (actual time=0.005..0.636 rows=201 loops=1)
         ->  Hash  (cost=2.00..2.00 rows=100 width=56) (actual time=1.479..1.479 rows=201 loops=1)
               Buckets: 1024  Batches: 1  Memory Usage: 15kB
               ->  CTE Scan on inputs  (cost=0.00..2.00 rows=100 width=56) (actual time=0.067..1.056 rows=201 loops=1)
   ->  Index Scan using cookbooks_pkey on cookbooks c  (cost=0.00..0.37 rows=1 width=82) (actual time=0.003..0.005 rows=1 loops=40401)
         Index Cond: (id = v.cookbook_id)
         Filter: (org_id = 'bb996f532e8358779d236750526af7c8'::bpchar)
 Total runtime: 418.822 ms
(15 rows)
```

After this patch is applied, the query runs as such:

```
                                                                            QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Hash Join  (cost=36.80..43.85 rows=1 width=393) (actual time=2.462..6.652 rows=201 loops=1)
   Hash Cond: ((minimal_cv.name = inputs.name) AND (minimal_cv.major = inputs.major) AND (minimal_cv.minor = inputs.minor) AND (minimal_cv.patch = inputs.patch))
   CTE inputs
     ->  Function Scan on unnest  (cost=0.00..1.00 rows=100 width=56) (actual time=0.048..0.349 rows=201 loops=1)
   CTE minimal_cv
     ->  Hash Join  (cost=8.03..31.80 rows=201 width=602) (actual time=0.905..3.174 rows=201 loops=1)
           Hash Cond: (v.cookbook_id = c.id)
           ->  Seq Scan on cookbook_versions v  (cost=0.00..21.01 rows=201 width=528) (actual time=0.005..0.334 rows=201 loops=1)
           ->  Hash  (cost=5.51..5.51 rows=201 width=82) (actual time=0.851..0.851 rows=201 loops=1)
                 Buckets: 1024  Batches: 1  Memory Usage: 23kB
                 ->  Seq Scan on cookbooks c  (cost=0.00..5.51 rows=201 width=82) (actual time=0.014..0.440 rows=201 loops=1)
                       Filter: (org_id = 'bb996f532e8358779d236750526af7c8'::bpchar)
   ->  CTE Scan on minimal_cv  (cost=0.00..4.02 rows=201 width=417) (actual time=0.913..4.233 rows=201 loops=1)
   ->  Hash  (cost=2.00..2.00 rows=100 width=56) (actual time=1.501..1.501 rows=201 loops=1)
         Buckets: 1024  Batches: 1  Memory Usage: 15kB
         ->  CTE Scan on inputs  (cost=0.00..2.00 rows=100 width=56) (actual time=0.052..1.026 rows=201 loops=1)
 Total runtime: 7.085 ms[A
(17 rows)
```

On the best case scenario, where each cookbook has a unique version,
both queries return within 5-7ms on my laptop in the development VM.

This case isn't tested. The best case above includes one cookbook
version for each cookbook and the version numbers are unique across
the space. The case that isn't tested is the "normal" case, where each
cookbook has multiple cookbook versions and they could be overlapping
with other cookbooks. Before this patch, there would be the potential to
"cross polinate" (my term) the join of `inputs` and `cookbook_versions`,
where we want the result for cookbook `foo` at `1.2.3` but cookbooks
`bar`,`baz`,and `bat` are also available at that version. We're still
going to join their data before filtering it by name in the next step.
Given the above, I'm strongly confident that this patch improves the
normal case for this query as well.